--- hhplus-server-java/src/main/java/kr/hhplus/be/server/infra/payment/JpaPaymentRepository.java
@@ -1,11 +1,0 @@
-package kr.hhplus.be.server.infra.payment;
-
-import kr.hhplus.be.server.domain.payment.entity.Payment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface JpaPaymentRepository extends JpaRepository<Payment, Long> {
-
-    Optional<Payment> findByOrderId(Long orderId);
-}
