--- hhplus-server-java/src/main/java/kr/hhplus/be/server/infra/payment/PaymentRepositoryImpl.java
@@ -1,25 +1,0 @@
-package kr.hhplus.be.server.infra.payment;
-
-import kr.hhplus.be.server.domain.payment.PaymentRepository;
-import kr.hhplus.be.server.domain.payment.entity.Payment;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-@RequiredArgsConstructor
-public class PaymentRepositoryImpl implements PaymentRepository {
-
-    private final JpaPaymentRepository jpaPaymentRepository;
-
-    @Override
-    public void save(Payment payment) {
-        jpaPaymentRepository.save(payment);
-    }
-
-    @Override
-    public Optional<Payment> findByOrderId(Long orderId) {
-        return jpaPaymentRepository.findByOrderId(orderId);
-    }
-}
