+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/infra/repository/order/JpaOrderAddressRepository.java
@@ -1,0 +1,14 @@
+package kr.hhplus.be.server.infra.repository.order;
+
+import kr.hhplus.be.server.domain.order.entity.OrderAddress;
+import lombok.NonNull;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface JpaOrderAddressRepository extends JpaRepository<OrderAddress, Long> {
+
+    @NonNull
+    Optional<OrderAddress> findByOrderId(@NonNull Long orderId);
+
+}
