+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/infra/repository/order/JpaOrderCouponRepository.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.infra.repository.order;
+
+import kr.hhplus.be.server.domain.order.entity.OrderCoupon;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaOrderCouponRepository extends JpaRepository<OrderCoupon, Long> {
+
+    OrderCoupon findByOrderId(Long orderId);
+
+}
