+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/infra/repository/order/JpaOrderHistoryRepository.java
@@ -1,0 +1,12 @@
+package kr.hhplus.be.server.infra.repository.order;
+
+import kr.hhplus.be.server.domain.order.entity.OrderHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface JpaOrderHistoryRepository extends JpaRepository<OrderHistory, Long> {
+
+    Optional<OrderHistory> findByOrderId(Long id);
+
+}
