+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/infra/repository/user/JpaUserRepository.java
@@ -1,0 +1,12 @@
+package kr.hhplus.be.server.infra.repository.user;
+
+import kr.hhplus.be.server.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface JpaUserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findById(long id);
+
+}
